Added start of logic backend
rando/Check adds the ability to make a list of checks with a regular expression as the requirement
rando/Input has been updated to include a regular expression for requirements
Randomize changes have been started
Requirement_list.txt is a personal note used for creating the regular expressions for each Check
The randomizer currently works the exact same as before. This update only contains the very start of the backend.